actions: split between the action description and a action-defined-by-a-method action
This removes the confusion from other type of actions (e.g. Syskit
action-from-requirement) which are not necessarily attached to an
action interface.
