+++ lib/roby/actions.rb
@@ -5,6 +5,7 @@ module Roby
 end
 
 require 'roby/actions/models/action'
+require 'roby/actions/models/method_action'
 require 'roby/actions/models/interface'
 require 'roby/actions/models/library'
