+++ lib/roby/actions/models/action.rb
@@ -22,9 +22,6 @@ module Roby
                 end
             end
 
-            # The action interface on which this action is defined
-            attr_accessor :action_interface_model
-
             # The action name
             attr_accessor :name
             # The action description
@@ -42,16 +39,20 @@ module Roby
             #
             # @return [Model<Roby::Task>,Model<Roby::TaskService>]
             attr_reader :returned_type
-            # If this action is actually a coordination model, returns it
-            #
-            # @return [nil,Coordination::Models::Base]
-            attr_accessor :coordination_model
 
             # @return [Action] an action using this action model and the given
             #   arguments
             def new(arguments = Hash.new)
                 Actions::Action.new(self, normalize_arguments(arguments))
             end
+
+            def ==(other)
+                other.kind_of?(self.class) &&
+                    other.name == name &&
+                    other.arguments == arguments &&
+                    other.returned_type == returned_type
+            end
+
 
             # Task model that can be used to represent this action in a plan
             def returned_task_type
@@ -69,8 +70,7 @@ module Roby
                 end
             end
 
-            def initialize(action_interface_model = nil, doc = nil)
-                @action_interface_model = action_interface_model
+            def initialize(doc = nil)
                 @doc = doc
                 @arguments = []
                 @returned_type = Roby::Task
@@ -80,12 +80,6 @@ module Roby
             def initialize_copy(old)
                 super
                 @arguments = arguments.map(&:dup)
-            end
-            
-            def ==(other)
-                other.kind_of?(self.class) &&
-                    other.action_interface_model == action_interface_model &&
-                    other.name == name
             end
 
             # @api private
@@ -125,6 +119,7 @@ module Roby
                 arguments << Argument.new(name.to_s, doc, true)
                 self
             end
+
             # Documents a new optional argument to the method
             def optional_arg(name, doc = nil, default = nil)
                 arg = Argument.new(name.to_s, doc, false)
@@ -175,101 +170,8 @@ module Roby
                 Kernel.validate_options arguments, self.arguments.map(&:name)
             end
 
-            # Instanciate this action on the given plan
-            def instanciate(plan, arguments = Hash.new)
-                run(action_interface_model.new(plan), arguments)
-            end
-
-            # Executes the action on the given action interface
-            def run(action_interface, arguments = Hash.new)
-                if self.arguments.empty?
-                    if !arguments.empty?
-                        raise ArgumentError, "#{name} expects no arguments, but #{arguments.size} are given"
-                    end
-                    result = action_interface.send(name).as_plan
-                else
-                    default_arguments = self.arguments.inject(Hash.new) do |h, arg|
-                        h[arg.name] = arg.default
-                        h
-                    end
-                    arguments = Kernel.validate_options arguments, default_arguments
-                    self.arguments.each do |arg|
-                        if arg.required && !arguments.has_key?(arg.name.to_sym)
-                            raise ArgumentError, "required argument #{arg.name} not given to #{name}"
-                        end
-                    end
-                    result = action_interface.send(name, arguments).as_plan
-                end
-                # Make the planning task inherit the model/argument flags
-                if planning_task = result.planning_task
-                    if planning_task.respond_to?(:action_model=)
-                        planning_task.action_model ||= self
-                    end
-                    if planning_task.respond_to?(:action_arguments=)
-                        result.planning_task.action_arguments ||= arguments
-                    end
-                end
-                result
-            end
-
-            # Create a new action model that is bound to a different interface model
-            #
-            # @param [Models::Interface] action_interface_model the new model
-            # @param [Boolean] force the rebind will happen only if the new
-            #   interface model is a submodel of the current one. If force is
-            #   true, it will be done regardless.
-            # @return [Action] the rebound action model
-            def rebind(action_interface_model, force: false)
-                m = dup
-                # We rebind only if the new interface model is a submodel of the
-                # old one
-                if force || (action_interface_model <= self.action_interface_model)
-                    m.action_interface_model = action_interface_model
-                    if coordination_model
-                        m.coordination_model = coordination_model.rebind(action_interface_model)
-                    end
-                end
-                m
-            end
-
-            # Returns the plan pattern that will deploy this action on the plan
-            def plan_pattern(arguments = Hash.new)
-                job_id, arguments = Kernel.filter_options arguments, :job_id
-
-                planner = Roby::Actions::Task.new(
-                    Hash[action_interface_model: action_interface_model,
-                    action_model: self,
-                    action_arguments: arguments].merge(job_id))
-                planner.planned_task
-            end
-
             def as_plan(arguments = Hash.new)
                 plan_pattern(arguments)
-            end
-
-            def proxy(peer)
-                interface_model = action_interface_model.proxy(peer)
-                if interface_model.respond_to?(:find_action_by_name) && (action = interface_model.find_action_by_name(name))
-                    return action
-                end
-
-                result = self.dup
-                result.proxy!(peer, interface_model, arguments)
-                result
-            end
-
-            def proxy!(peer, interface_model, arguments)
-                @action_interface_model = interface_model
-                @returned_type = returned_type.proxy(peer)
-                @arguments = arguments.proxy(peer)
-            end
-
-            def to_s
-                if action_interface_model
-                    "#{action_interface_model.name}.#{name}"
-                else
-                    "<anonymous>.#{name}"
-                end
             end
 
             def pretty_print(pp)
@@ -292,16 +194,15 @@ module Roby
                 end
             end
 
-            # Returns the underlying coordination model
-            #
-            # @raise [ArgumentError] if this action is not defined by a
-            #   coordination model
-            # @return [Model<Coordination::Base>]
-            def to_coordination_model
-                if coordination_model
-                    coordination_model
-                else raise ArgumentError, "#{self} does not seem to be based on a coordination model"
-                end
+            def proxy(peer)
+                result = self.dup
+                result.proxy!(peer, arguments)
+                result
+            end
+
+            def proxy!(peer, arguments)
+                @returned_type = returned_type.proxy(peer)
+                @arguments = arguments.proxy(peer)
             end
 
             def to_action_model
